WebSockets/tests/submissions/Opera/interfaces/WebSocket/readyState/004.html

<!doctype html>
<title>WebSockets: defineProperty getter for readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  assert_throws(new TypeError(), function(){
    Object.defineProperty(WebSocket.prototype, 'readyState', {
      get: function() { return 'foo'; }
    });
  });
});
</script>
